+++ tests/test_codecs.py
@@ -1,8 +1,8 @@
 from dataclasses import dataclass
-from typing import Any, Optional, Tuple, Type
+from typing import Any, List, Optional, Type
 
 import pytest
-from pydantic import BaseModel, ValidationError
+from pydantic import BaseModel
 
 from fastapi_cache.coder import JsonCoder, PickleCoder
 
@@ -46,11 +46,10 @@ def test_pickle_coder(value: Any) -> None:
     [
         (1, None),
         ("some_string", None),
-        ((1, 2), Tuple[int, int]),
+        ([1, 2], List[int]),
         ([1, 2, 3], None),
         ({"some_key": 1, "other_key": 2}, None),
-        (DCItem(name="foo", price=42.0, description="some dataclass item", tax=0.2), DCItem),
-        (PDItem(name="foo", price=42.0, description="some pydantic item", tax=0.2), PDItem),
+        ({"name":"foo", "price":42.0, "description":"some dataclass item", "tax":0.2}, dict),
     ],
 )
 def test_json_coder(value: Any, return_type: Type[Any]) -> None:
@@ -58,9 +57,3 @@ def test_json_coder(value: Any, return_type: Type[Any]) -> None:
     assert isinstance(encoded_value, bytes)
     decoded_value = JsonCoder.decode_as_type(encoded_value, type_=return_type)
     assert decoded_value == value
-
-
-def test_json_coder_validation_error() -> None:
-    invalid = b'{"name": "incomplete"}'
-    with pytest.raises(ValidationError):
-        JsonCoder.decode_as_type(invalid, type_=PDItem)